[BUGFIX] Separate middleware initialization
The frontend usergroup needs to be set before the typo3/cms-frontend/page-resolver is processed. At that time the TSFE controller isn't initialized yet, so the TSFE configuration needs to be done at a later point.
